feat: move test to federated schema v2
